requirements.txt
python-kis>=2.1.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
ta>=0.11.0
APScheduler>=3.10.0
python-telegram-bot>=21.0
aiosqlite>=0.20.0

# AI strategy
google-generativeai>=0.8.0
anthropic>=0.40.0
openai>=1.50.0
